LPS-56703 Move DisplayContexts to OSGi
