+++ modules/apps/document-library/document-library-web/src/com/liferay/document/library/web/display/context/DLDisplayContextFactoryReference.java
@@ -12,9 +12,10 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.display.context;
+package com.liferay.document.library.web.display.context;
 
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portlet.documentlibrary.display.context.DLDisplayContextFactory;
 import com.liferay.registry.ServiceReference;
 
 /**
