+++ modules/apps/document-library/document-library-web/src/com/liferay/document/library/web/display/context/DLDisplayContextProvider.java
@@ -12,12 +12,14 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.display.context;
+package com.liferay.document.library.web.display.context;
 
 import com.liferay.portal.kernel.display.context.DisplayContextProvider;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.repository.model.FileShortcut;
 import com.liferay.portal.kernel.repository.model.FileVersion;
+import com.liferay.portlet.documentlibrary.display.context.DLEditFileEntryDisplayContext;
+import com.liferay.portlet.documentlibrary.display.context.DLViewFileVersionDisplayContext;
 import com.liferay.portlet.documentlibrary.model.DLFileEntryType;
 
 import javax.servlet.http.HttpServletRequest;
