+++ modules/apps/document-library/document-library-web/src/com/liferay/document/library/web/display/context/IGDisplayContextProviderImpl.java
@@ -12,13 +12,15 @@
  * details.
  */
 
-package com.liferay.portlet.imagegallerydisplay.display.context;
+package com.liferay.document.library.web.display.context;
 
 import com.liferay.portal.kernel.display.context.BaseDisplayContextProvider;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.repository.model.FileShortcut;
 import com.liferay.portal.kernel.repository.model.FileVersion;
+import com.liferay.portlet.imagegallerydisplay.display.context.IGDisplayContextFactory;
+import com.liferay.portlet.imagegallerydisplay.display.context.IGViewFileVersionDisplayContext;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
