+++ modules/apps/document-library/document-library-web/src/com/liferay/document/library/web/display/context/dependencies/open_in_ms_office_js.ftl
