+++ modules/apps/document-library/document-library-web/src/com/liferay/document/library/web/display/context/logic/DLPortletInstanceSettingsHelper.java
@@ -12,8 +12,10 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.display.context.logic;
+package com.liferay.document.library.web.display.context.logic;
 
+import com.liferay.document.library.web.display.context.util.DLRequestHelper;
+import com.liferay.document.library.web.settings.internal.DLPortletInstanceSettings;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.KeyValuePair;
@@ -24,8 +26,6 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PortletKeys;
 import com.liferay.portal.util.PropsValues;
-import com.liferay.portlet.documentlibrary.DLPortletInstanceSettings;
-import com.liferay.portlet.documentlibrary.display.context.util.DLRequestHelper;
 import com.liferay.portlet.documentlibrary.util.DLUtil;
 
 import java.util.ArrayList;
