+++ modules/apps/document-library/document-library-web/src/com/liferay/document/library/web/display/context/util/JSPRenderer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.display.context.util;
+package com.liferay.document.library.web.display.context.util;
 
 import java.io.IOException;
